Modified test code to use new test databse
Modified code to accept the connection string as a parameter to DAL methods and wrote test code that passes in the connection string for the test database. Wrote tests for the IngredientDAL and RecipeDAL classes in the RecipePlannerLibrary project to ensure the functionality works as expected. Achieved 100% test coverage and all tests passed for these two classes.