Modified search tags functionality
Modified functionality so that user can filter by more than one tag and tags that are being used to filter are displayed below the search bar. This new functionality allows users to narrow down the options as much as they need and find recipes that better suit their needs at the time.