Fixed bug with adding ingredient
Bug was causing duplicate ingredients with different id numbers to be added to the ingrerdient info table. Updated the insert sql statements so that ingredient is added to the ingredient info and ingredient tables with a new id if it does not already exist other wise we get the ingredient id from the ingredient info table and use that to add the ingredient to the ingredient table. Fixed failing tests. Added updated scripts