Added Confirmation for Cook Meal
Added confirmation page that notifies user that ingredients will be removed from their pantry and asks if they'd like to proceed. If cancel button is pressed, it takes the user back to the page they came from. If continue button is pressed the ingredients are removed and they are routed to the pantry.