Fixed Add Ingredient bug
Fixed bug that prevented users from adding an ingredient to the shopping list if it was already in the pantry. This has been resolved and user can now add ingredient to shopping list if it is already in the pantry but not if it is already in the shopping list. They will need to update quantity instead.