Added verification to share recipe functionaly
Previously user was able to enter a username and press share without selecting a recipe which casued the program to crash. Added verification that notifies user that recipe has to be selected before they can share to the given username. Error is no longer thrown and system runs functionality as expected