Modified the RecipeStepAdd page functionality for the up/down columns and figured out updating the steps column when a row is removed.
Changed how the up/down columns are handled. Also switched the up/down columns to be before the remove button in the list view. Figured out keeping the sequential order when a row is removed from the list view. Finally, added method and class specifications to the classes that were missing specifications.